+++ gobot_software/include/gobot_software/gobot_status.hpp
@@ -24,6 +24,8 @@ std::string getUpdateStatus();
 
 void updateStatus(std::string &str);
 
+void threadVoice(std::string file, int mute);
+
 void robotResponse(int status, std::string text);
 
 void setHomePose(void);